memdup int to size_t

+++ include/shared.h
@@ -19,6 +19,6 @@ void swap(void *a, void *b, const size_t size);
 size_t getFileSize(FILE *fp);
 int getFmtSize(const char *fmt, ...);
 int getFmtSizeVa(const char *fmt, va_list ap);
-void *memdup(const void *mem, const int size);
+void *memdup(const void *mem, const size_t size);
 
 #endif
